TECH: Refactor store utils
