--- src/react/store/combineReducers.js
@@ -1,10 +1,0 @@
-const combineReducers = reducers => (state, action) =>
-  Object.entries(reducers).reduce(
-    (obj, [key, reducer]) => ({
-      ...obj,
-      [key]: reducer(state[key], action),
-    }),
-    {}
-  )
-
-export default combineReducers
