--- src/react/store/createReducer.js
@@ -1,8 +1,0 @@
-const createReducer = handlers => (state, action) => {
-  if (!handlers.hasOwnProperty(action.type)) {
-    return state
-  }
-  return handlers[action.type](state, action)
-}
-
-export default createReducer
