+++ src/react/store/index.js
@@ -1,8 +1,22 @@
 import reducers from './reducers'
-import combineReducers from './combineReducers'
+
+const combineReducers = reducers => (state, action) =>
+  Object.entries(reducers).reduce(
+    (obj, [key, reducer]) => ({
+      ...obj,
+      [key]: reducer(state[key], action),
+    }),
+    {}
+  )
 
 export const reducer = combineReducers(reducers)
 
 export { default as initialState } from './initialState'
 
 export { StoreProvider, useStore, useDispatch } from './StoreContext'
+export {
+  useDispatchAction,
+  usePromiseDispatch,
+  createReducer,
+  promiseReducers,
+} from './storeUtils'
