+++ src/react/utils/index.js
@@ -1,3 +1,1 @@
 export { debounce } from './debounce'
-export { debounceDispatchActionPromise } from './debounceDispatchActionPromise'
-export { promiseActionReducer } from './promiseActionReducer'
